Dorfman Group Testing Decoder

Implements two-stage Dorfman group testing algorithm with sensitivity and specificity considerations.

Usage

dorfman(Y.true, Se, Sp, c)

Arguments

Y.true

Numeric vector (n x 1) of true binary statuses (1=positive, 0=negative)

Se

Numeric vector of length 2: (group sensitivity, individual sensitivity)

Sp

Numeric vector of length 2: (group specificity, individual specificity)

c

Integer group size (must divide length(Y.true))

Value

List with:

Z

Test results matrix (J x (c+3)) where columns are: [Test result, Group size, Test type (1=group, 2=individual), Sample indices]

Y

Individual results matrix (n x 4) where columns are: [Test result, Test type, Group test ID, Individual test ID]

Details

The function performs:

Group testing: Samples are pooled into groups of size c and tested

Individual testing: If a group tests positive, all members are tested individually

The probability models used: $$P(Z_g = 1|Y_g) = \begin{cases} Se[1] & \text{if } \sum Y_g > 0 \\ 1-Sp[1] & \text{otherwise} \end{cases}$$ $$P(Z_i = 1|Y_i) = \begin{cases} Se[2] & \text{if } Y_i = 1 \\ 1-Sp[2] & \text{otherwise} \end{cases}$$ where \(Z_g\) is group test result, \(Z_i\) is individual test result, and \(Y_g\), \(Y_i\) are true statuses.

Examples

set.seed(123)
Y.true <- rbinom(10, 1, 0.3)
result <- dorfman(Y.true, c(0.95, 0.98), c(0.98, 0.99), 2)
result
#> $Z
#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]    0    2    1    1    2
#>  [2,]    1    2    1    3    4
#>  [3,]    0    1    2    3  -99
#>  [4,]    1    1    2    4  -99
#>  [5,]    1    2    1    5    6
#>  [6,]    1    1    2    5  -99
#>  [7,]    0    1    2    6  -99
#>  [8,]    1    2    1    7    8
#>  [9,]    0    1    2    7  -99
#> [10,]    1    1    2    8  -99
#> [11,]    0    2    1    9   10
#> 
#> $Y
#>       [,1] [,2] [,3] [,4]
#>  [1,]    0    1    1  -99
#>  [2,]    0    1    1  -99
#>  [3,]    0    2    2    3
#>  [4,]    1    2    2    4
#>  [5,]    1    2    5    6
#>  [6,]    0    2    5    7
#>  [7,]    0    2    8    9
#>  [8,]    1    2    8   10
#>  [9,]    0    1   11  -99
#> [10,]    0    1   11  -99
#>