gtm: Modeling Group Testing Data

The gtm package provides tools for modeling and analyzing group testing data. It includes functions for fitting models, evaluating performance, and visualizing results.

See also

Author

Maintainer: Muhammad Yaseen myaseen208@gmail.com (ORCID) (School of Mathematical & Statistical Sciences, Clemson University, Clemson, South Carolina, USA)

Authors:

Christopher McMahan mcmaha2@clemson.edu (ORCID) (School of Mathematical & Statistical Sciences, Clemson University, Clemson, South Carolina, USA)

Christopher Bilder bilder@unl.edu (ORCID) (Department of Statistics, University of Nebraska-Lincoln, Nebraska, USA)

Joshua Tebbs tebbs@stat.sc.edu (ORCID) (Department of Statistics, University of South Carolina, Columbia, South Carolina, USA)